feat(card-management): card generator load tests

+++ services/card-management/src/main/java/com/processing/cardmanagement/options/CardServiceDefaults.java
@@ -8,7 +8,7 @@ public interface CardServiceDefaults {
     /**
      * @return значение лимита для пагинации
      */
-    long pageLimit();
+    int pageLimit();
 
     /**
      * @return значение сдвига для пагинации
