+++ services/card-management/src/main/java/com/processing/cardmanagement/options/CardServiceDefaultsConfigurationProperties.java
@@ -12,7 +12,7 @@ import org.springframework.validation.annotation.Validated;
 public record CardServiceDefaultsConfigurationProperties(
 
     @Positive
-    long pageLimit,
+    int pageLimit,
 
     @NotNegative
     long pageOffset,
